+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version `v3.0.6` (13.10.2025)
 * Support "delete" operation in secure store (FSSP-1)
+* SecretStore implementation for folio-secure-store-proxy (APPPOCTOOL-59)
 
 ---